ci(1555): shard services-api across three runners, and stop the vault tests sleeping
The services-api job took ~8.5 min: ~6 min of pytest on a 4-vCPU runner that
`-n auto` cannot widen, most of it building the FastAPI app ~1000 times.

- `--shard k/N` moves from the integration conftest to the root one, so any
  tier can use it; services-api now runs as three shards (1389/1390/1389 tests,
  split by file after collection, so a new file cannot be missed)
- the vault client's failure-path tests waited through real retry backoff, 7s
  each; they now retry without waiting (46 tests: ~56s -> 0.3s)
