+++ services/tests/integration/conftest.py
@@ -23,8 +23,6 @@ from terrapod.api.dependencies import (
 )
 from terrapod.db.models import Base
 
-from ..meta.shard_plan import plan_shards
-
 # ---------------------------------------------------------------------------
 # Ensure test-friendly defaults (must precede any Settings import)
 # ---------------------------------------------------------------------------
@@ -342,61 +340,3 @@ async def assign_platform_role(engine, provider: str, email: str, role_name: str
             ),
             {"p": provider, "e": email, "r": role_name},
         )
-
-
-# ── Runner-level sharding (#1468) ─────────────────────────
-#
-# `--shard k/N` keeps only the files assigned to shard k. Splitting happens
-# AFTER collection, so it sees the true set of files and a new test file cannot
-# be silently missed — the failure mode of a hand-maintained path list.
-#
-# Runner-level, not xdist: the objection to xdist here is that its workers share
-# one database and race to create the same tables. Separate runners each get
-# their own compose stack, which is how the unit and services-api slices already
-# work.
-
-
-def pytest_addoption(parser):
-    parser.addoption(
-        "--shard",
-        default=None,
-        help="Run only this shard of the suite, as k/N (1-based). Splits by file, "
-        "balanced on collected test count.",
-    )
-
-
-def pytest_collection_modifyitems(config, items):
-    spec = config.getoption("--shard")
-    if not spec:
-        return
-
-    try:
-        index_s, total_s = spec.split("/", 1)
-        index, total = int(index_s), int(total_s)
-    except ValueError:
-        raise pytest.UsageError(f"--shard expects k/N, got {spec!r}") from None
-    if not 1 <= index <= total:
-        raise pytest.UsageError(f"--shard {spec}: k must be within 1..N")
-
-    counts: dict[str, int] = {}
-    for item in items:
-        counts[str(item.path)] = counts.get(str(item.path), 0) + 1
-
-    keep = set(plan_shards(counts, total)[index - 1])
-    selected = [i for i in items if str(i.path) in keep]
-    deselected = [i for i in items if str(i.path) not in keep]
-
-    # Printed so a shard that selects nothing is visible in the log rather than
-    # passing as a vacuous success.
-    print(
-        f"\nshard {index}/{total}: {len(keep)} files, {len(selected)} tests "
-        f"({len(deselected)} deselected)"
-    )
-    if not selected:
-        raise pytest.UsageError(
-            f"--shard {spec} selected no tests. With {len(counts)} files collected "
-            "this means the split is wrong, not that there is nothing to run."
-        )
-
-    config.hook.pytest_deselected(items=deselected)
-    items[:] = selected
